+++ apps/api-docs/CHANGELOG.md
@@ -22,6 +22,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.6.0] - 2023-05-18
+
+### Added
+
+-   `308` response for `getNodes`.
+
+## [2.5.1] - 2023-05-14
+
+### Changed
+
+-   Updated API version from `2.5.0` to `2.5.1`.
+
+## [2.5.0] - 2023-05-03
+
+### Changed
+
+-   Revised response types for `getResolveObjects` and `postResolveObjects`.
+
 ## [2.4.0] - 2023-04-25
 
 ### Added