+++ packages/api-models/CHANGELOG.md
@@ -21,6 +21,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [1.3.2] - 2023-05-14
+
+### Changed
+
+-   Upgraded to TypeScript 5.0.4.
+
 ## [1.3.1] - 2023-04-30
 
 ### Fixed